The journey to doing more with less
PumpkinDB is my forth public attempt to perfect my vision for an event sourcing
based database.

I started off with some private prototypes and, eventually,
[Genomu](https://github.com/genomu/genomu) a few years ago. It was written in
Elixir, had a really bastardized version of event
sourcing and didn't go far. But I did a presentation about it:

https://www.youtube.com/watch?v=ujaoprIKY9M.

Fast-forward to 2016, I still can't let these ideas go. I've implemented a few
simple prototypes in clients' projects and then developed
[eventreducer](https://github.com/eventreducer/eventreducer) in an attempt to create
a more generic solution.

The main idea that was haunting me around that time was that instead of
processing events into some domain models we can simply index events and
build (plural) domain models on the fly by querying those. It is still very much
a central idea in the target design I am pursuing in PumpkinDB.

But I didn't like the implementation and some choices I made, so I rewrote it.
This is how [es4j](https://github.com/eventsourcing/es4j) was started (although
originally it was called eventchain -- you can tell I am focusing on naming quite a lot).

To date, es4j (even with all its warts, of which it has an abundance of) was the
most complete and sophisticated implementation I did so far. I've learned tons and used
it in real projects. I spoke about it at an O'Reilly event: https://www.youtube.com/watch?v=aqv8d1pjmU8

But as you might have guessed already, this wasn't perfect enough, either.

Firsly, es4j was overly restrictive as what an event (or a command) should be. It had to have
a UUID, an [HLC](https://www.cse.buffalo.edu/tech-reports/2014-04.pdf) timestamp and
had to conform to [ELF format](https://rfc.eventsourcing.com/spec:1/ELF) and its encodings
which had no users besides me and a handful of people who used es4j. While not a problem
for me (I still enjoy ELF and [BEST](https://rfc.eventsourcing.com/spec:2/BEST) a lot), it
unnecessarily raised a barrier where it shouldn't have been. It should have allowed for more flexibility
in structure, typing and encoding.

Thus, PumpkinDB lifts all of these restrictions and treats events, commands and
their identifiers as binary blobs.

How does it allow indexing to happen if it doesn't know anything about the
event's encoding or format?

This is where another big piece of PumpkinDB's core design comes in. One of the
central pieces of it is a small [Forth-like
language](https://en.wikipedia.org/wiki/Forth_(programming_language)).

This is the most low-level way to communicate with the database. Every
transaction is in fact a script that journals and indexes the data being
recorded. And because PumpkinDB provides primitives for different
encodings (JSON, ELF/BEST, XML, etc.) you can be very flexible in
how you're processing the data.

One level higher, PumpkinDB provides practical generalized scripts to accomplish most
common scenarios to make the communication with it easier and more compact.

What is interesting about this language is that it will also enable writing
manual data-iterating queries over indices and journal to accomplish any
possible query. But since in most cases it will be an overkill and a barrier to entry,
PumpkinDB will provide a higher-level declarative query language that'll cover most
of the needs, so the manual querying will be left to those who need
to exercise control over how the query is done.

Secondly, es4j was very much limited to Java. Yes, a server component could
have been written, but it never was. It would have required some heavy
refactoring in some places. So it was a barrier.

PumpkinDB is being developed in Rust. This doesn't, however, mean that one
would need to write Rust to use it. This time I am going back to a thin client
design to enable developers in different languages use it without much of a difficulty.

Why Rust? I wanted something with the least amount of overhead and a good sense
of safety when it comes to a lower-level language.

Thirdly, es4j was starting outgrowing its pluggable backend mechanism in terms
of cooperative abilities (optimizations and proper transactions) and the indexing
engine that it was based on ([CQEngine](https://github.com/npgall/cqengine), read
more about CQEngine-related issues here: https://github.com/eventsourcing/es4j/issues/173).

PumpkinDB is using a single shared backend (LMDB at the moment) to take advantage of
optimizations and perfect transaction isolation across journal and indices.

es4j is still being maintained as PumpkinDB is being developed (existing
projects rely on it, after all) but I believe that, going forward, the ideas in
PumpkinDB are much simpler and allow to achieve a lot more with a lot less.

---

Lastly, but not least importantly, while I've referred a lot to myself and my
own experiences in this write-up, just like es4j, this project is meant to be
driven collaboratively. I am still a huge fan of ideas promoted by
[Pieter Hintjens](https://en.wikipedia.org/wiki/Pieter_Hintjens),
particularly those codified in [C4](https://rfc.unprotocols.org/spec:1/C4).

In a nutshell, this means:

* We merge pull requests rapidly (try!)
* We are open to diverse ideas
* We prefer code now over consensus later

One slight deviation from the C4 process is to allow one more type of commit
messages (other than "Problem: ... Solution: ..."). It is *this* type of commit
messages -- pure write-ups, coming with no changes to files.

The motivation for this is that web is not a reliable place to retain articles at
(hosts go down, content gets deleted, etc.). Nor it's easy to find relevant pieces
with all the noise out there.

What did the contributor think about when he was developing this or that
part? What train of thought was he on?

Keeping the articles in the git log allows to retain them forever (for as long
as there's at least one copy of the repository somewhere) and provide
context to those who really want to learn more about the project.

---

Why was it named PumpkinDB? Firstly, it is a reference to one of the articles I
wrote on the subject (see a72e9bb). Secondly, it's the result of the [popular
vote on Twitter](https://twitter.com/yrashk/status/826648287713255424):

```
[ 9% ] LESS (Lazy Event Src Srv)
[      17% ] BLESS (B = ?)
[ 9% ] Causality
[                                          65% ] Pumpkin
````

Hope that helps!