Problem: how do we make PumpkinDB really flexible?
In previous incarnations (or, rather, inspirations) of PumpkinDB much more
rigid structures, formats and encoding were established as a prerequisite for
using it, unnecessarily limiting the applicability and appeal of the technology
and ideas behind it. For example, one had to buy into
[ELF](https://rfc.eventsourcing.com/spec:1/ELF), UUID-based event
identification and [HLC-based](https://rfc.eventsourcing.com/spec:6/HLC)
timestamps.

Solution: PumpkinDB's operational foundation should be a simple and powerful
programming language so that it can be used to describe all possible nuances.

PumpkinScript is a minimalistic concatenative, stack-based language inspired by
Forth.

It is used in PumpkinDB to operate a low-level database "virtual machine" — to
manipulate, record and retrieve data.

The idea is that end users will rarely or never see this language when they
work with PumpkinDB.

This commit lays the very foundation of PumpkinScript. It's incomplete but we
need to start somewhere. Currently it only supports pushing data on the stack
and a very basic set of stack words.