Problem: using slice references in script stacks is problematic
First of all, this is an immediate concern for forming scripts
out of some kind of inputs. Something that should be as simple as

```rust
let program = script::parse(input);
for instruction in program {
	vm.execute(&mut stack, instruction).wait().unwrap();
}
```

is pretty much impossible because if we take a slice of that instruction
(and it's a Vec) it will be too short-lived to put it into the stack.

Solution: refactor script to use Vec<u8> instead of &[u8]

This does mean a lot more copying, of course, but so far I don't
see any other viable options.

Also, fixed a bug in the text-based script parser that prevented
parsing of words that were at the end of the input.