Problem: script composition is possible only over stack
However, there will be cases where being able to wrap fragments of code and
call them from other programs is highly beneficial.

For example, being able to wrap a piece of code into a transaction. Or, even
simpler, run a loop.

Solution: introduce EVAL word that will take topmost value off the stack
and evaluate it as a program, using current stack.

Also, add [] syntax to the parser to allow instructions like this:

```
0x10 [DUP DUP] EVAL
```