Problem: PumpkinScript interpreter isn't zero-copy
I mean, it's not even trying. It's operating on vectors exclusively
and thus can't provide important functionality of being able to
put data in the storage on the stack without copying it first.

Another problem is that with futures, the code is really messy
and hard to follow.

Solution: switch to a pre-allocated stack array of references,
an internal heap for allocation.

Also, get rid of futures, while we are at it. Futures have been
replaced with PumpkinScript's own scheduler that communicates
with the external words using queues (thanks to multiqueue).

This is a significant overhaul of the implementation.