Problem: using multiqueue is problematic
`multiqueue::BroadcastSender` does not implement
Sync` which is required by some packages (like Tokio),
preventing from passing the channels through closures.

Solution: switch to a much simpler, stabler (and slower)
solution -- Rust's channels.

The search for a better queue continues. I've heard about
some really cool work done in that space by a friend of mine,
but it's not available yet.

In the hindsight, the Sync problem was still there with
mpsc Sender, but was solved with the use of Mutex<Sender<...>>.

That said, the notification model used by the previous implementation
was unnecessarily complex. This change straightens this out. Once
the best queue is found, it can be relatively easily switched to
(or at least that's the hope).