Problem: database without a storage is not a database
That's just reality.

Solution: introduce first storage-related constructs into
PumpkinScript.

First script to work is:

```
["Hello" "world" ASSOC COMMIT] WRITE ["Hello" RETR] READ
```

This means that WRITE word will evaluate the first closure,
which associates key `"Hello"` with value `"world"` and
commits it, then the READ will evaluate the second closure,
which retrieves the value associated with the key `"Hello"`.

The reason for this model is that everything has to be
wrapped into transactions (that's how LMDB works anyway)

Also, rehaul some of the internals of the VM. For example,
that ugly HashMap tracking Envs is gone.

The result of this effort is not without some (temporary?)
disappointments. There was a realization that `Env<'a>`
wants all referenced elements on the stack to allow have
the same lifetime of `'a`. However, the lifetime of the
values extracted through lmdb is limited by the lifetime
of the transaction.

This means that right now we have to resort to copying
values. However, I suspect there's still a chance we can
have an optimization here. We might be able to have a
temporary "transaction context" stack that has a different
lifetime. The feasibility and consequences of such an idea
are still to be researched. But if defined well, this
should definitely improve the overall quiality and performance
of PumpkinDB.

But we've got to at least get a proof of concept out, so
copying these values is okay for now. Hopefully.

P.S. Oh, and for now we're using a fork of the lmdb-zero
crate. I am not sure if my patch is correct, but at least
it made it all work. See https://github.com/AltSysrq/lmdb-zero/pull/5
for more details. May be a better solution will come out of this?