Problem: by default, lmdb map size is small: 10485760
This effectively limits the database to approx 10Mb in size

Solution: detect available disk space and use that as a map
size, making the database file virtually unlimited. Until
we hit more complicated scenarios with multiple files on
different disks (with merged traversals, etc.)

This is not enabled on Windows as setting this on Windows
leads to making the file assume the size of the map.