Problem: Travis CI build fails
The test process gets killed, presumably heap growth
test issue.

Solution: indeed, it was disovered that by accident,
the code was written in such a way that a lot of memory
was allocated (over 3.5G), which caused smaller VMs to
kill the process. Didn't happen on my Mac. I had more
memory available.

What's more important, this problem uncovered a more
serious issue: heap allocator didn't do an allocation
when the requested chunk was perfectly aligned with the
end of the current heap, thus providing a very unsafe
slice back to the requestor. This has been fixed.