Problem: keeping values on the stack or repeating code is not fun
Since it is rather bothersome to keep certain values (like handles
or strings) around by manipulating the stack, it'd be nice to be able
to refer to them directly.

Solution: introduce `SET` word

It's syntax is rather interesting:

```
[<word name> : ...code...] SET
```

`SET` allows to define a value of a word for the scope of the script's
remainder. Keep in mind that `SET` does not evaluate the expression
after <word name>, it simply stores. In effect, each time <word name>
is called, it's re-evaluated again.