Problem: persisting keys in binary script form is bad for traversing
Solution: make Env stack store references to data without their size prefixes
(as slices already have length) and hence we can move to persisting data as is
in the database. Having stack that was is not big of a deal because we can
write size prefixes on demand when we are sending data back, avoiding
allocations there.

Fixes #23