Problem: no ordering sequencer primitive
In event-sourced systems it often makes sense to
have a unique, monotonically growing sequencer for journalling.

Solution: implement a set of words for HLC (Hybrid Logical Clocks,
https://www.cse.buffalo.edu/tech-reports/2014-04.pdf)

It carries both wall clock timestamp and a logical counter and is guaranteed to
provide unique values every time.

Eventually we may (or may be even should) have other sequencers as well.