Problem: unable to express optionality of a value
Some words might need to push values that are optional. An
upcoming example of that would be cursor words. Cursor iterator
might produce a key value pair, but it also might not (no pairs
left)

Solution: introduce the concept of an optional "value set" which
is basically just a closure with non-word items.

To analyze the value, `NONE?` and `SOME?` words can be used, and
to extract the actual value(s), one can use `UNWRAP`, which will
put them on the top of the stack.