Problem: impossible to extract a portion of a byte array
This is important when (for example) comparing key prefixes

Solution: SLICE word that would take a byte array, start index and end index
and push back a slice of that array.

Also, it was discovered that an empty string syntax ("") was
incorrectly parsed to an empty vector, so this has been fixed as well.

Closes #28