Problem: SET syntax is not composable
The problem with SET syntax is that it is quite irregular (SET has to look into
the supplied closure) and therefore composes very poorly.

For example, one can't supply a value and a name separately to SET through the
stack. The values have to be wrapped in a very particular
way.

Solution: split SET into SET and DEF and, most importantly,
introduce composable syntax by leveraging newly introduced `'WORD`
text form syntax that encodes the word into a binary format.

```
CURSOR 'c SET
[EQUALS?] 'equals? DEF
```

Closes #33