Problem: text protocol is limited
It doesn't allow for streaming or compact communication.

Solution: use binary-frames protocol with binary form encoding
and write a CLI tool (pumpkindb-term) to be the client REPL.

Also, switched to Mio from Tokio as I think it might fit into our
patterns better.

In the process of making this change, this project
turned from an application to both an application and
a library (to be used by utility applications like pumpkindb-term)

Closes #40