Problem: error details are encoded incorrectly
Instead of [descr [value] code] it is encoded as
[descr value code]

Solution: always encode details as a closure.

In the future, we might extend the error_program! macro
to support multiple value arguments (didn't have a need yet,
but it will come one day)

An important consequence of this change is that the words
are no longer written into details the same way.

Previously, they were written as [WORD]. However, one must
remember than 'WORD and [WORD] are no different as they are
just encoded values.

They are now encoded as ['WORD], and this is actually right
because previous version would violate the "unwrappability"
of the values, while this won't:

```
PumpkinDB> ['DUP] UNWRAP
0x83445550
PumpkinDB> [DUP] UNWRAP
Error: "Invalid value" 0x83445550 0x03
```