Problem: interpreting data pushes involves copying
Every time we push data from the program into the stack,
we copy it into Env's heap. That's wasteful. We already
have the program to refer to.

Solution: switch to a stack of programs  so that we can
use simple slices to read data from. Now injection happens
not over a return value but its pushed into the stack of
things to execute.

This also helps us reducing copying when creating new
closures wrapping other closures, as we can now split
them into few chunks to be pushed onto the program
stack independently.