Problem: querying "the last update" isn't trivial
Since PumpkinDB doesn't allow to override keys, one has to write to a new key
value every time something is recorded.

We have at least one primitive for that right now, and that is HLC. Since it
is guaranteed to generate unique and monotonically growing values, they can be
used to sequence any collection of values. All one has to do is to compose the
key this way:

```
["key" HLC CONCAT "value" ASSOC COMMIT] WRITE
```

Similarly, things like journalling events can be done in the same way:

```
["journal" HLC CONCAT "id" "value" 2 WRAP ASSOC COMMIT] WRITE
```

However, how do we quickly find what's the last element in the collection?

Solution: introduce ?CURSOR/SEEKLAST that will always efficiently search
for the last key with a given prefix.

Closes #80 (at least for now!)