Problem: some of tests failing on Windows
Solution: turns out, Windows was a great tool discovering
an issue in how we return the stack back from VM. It was
returned with a lifetime of the VM (which by that time might
have been gone). Now, a Vec<Vec<u8>> is returned instead
(a copy of the stack) which can safely be used later.

It wasn't a problem during testing on Linux and macOS, but
turned out to be an issue on Windows. Thanks, Microsoft!