Problem: No way to check if a value is INT or UINT.
Solution: Add a `INT?` word.

Since we do not have a type tag that indicates the
actual type (signed or unsigned) this is not a
very safe implementation. For instance, `256 INT?`
will return `1`, even tho the input is not a
signed integer.

If we want to safely support words that check for
types, we need to introduce type tags.