Problem: Too many statics (`STORAGE`)
We use `lazy_static` as a clutch too much as it’s
not really recommended way of working in rust if
it is possible to prevent it.

Solution: Remove the `STORAGE` static as it is
not needed.

To do this, the thread needs to move a reference
to the storage into it’s context as it is spawned,
the previous design didn’t allow for this as
easily since the `sender` had to be returned to
spawning thread. To solve this a single-use
channel is used to communicated the newly created
sender to the spawning thread. This channel will
only be used for that one purpose afterwards no
references will exist pointing to it.