Problem: Too many statics (`HLC_CLOCK`)
We use `lazy_static` as a clutch too much as it’s
not really recommended way of working in rust if
it is possible to prevent it.

Solution: Remove the `HLC_CLOCK` static as it is
not needed.

The only way that I can see having spent
considerable time reading over Rust’s
documentation on sharing references safely between
threads is by using `Arc`, or a smart pointer that
uses atomic operations to count references.

This will add some memory overhead (the amount
escapes me, but it is similar to `Rc` or two
`usize` values) as well as atomic reference
counting overhead but those are only done during
cloning and when the last `Arc` goes out of scope.
In PDB’s situation the last `Arc` will only go out
of scope once the server is shut down. Also, we do
not dynamically send the timestamp between
threads, it only happens during startup for a
constant number of threads (that constant is not
the same between instances, tho).

As before, the `HLC` clock is protected with a
`Mutex`.