Problem: HLC guarantees can be broken by sys clock
The system clock can go backwards in multiple
cases, if it has skewed to far forward for
instance. This will break the HLC promises since
the HLC timestamp is generated using the system
clock.

Solution: Introduce a `scratchpad` which is a meta
database internal to PumpkinDB. This file is
memory mapped into memory and individual modules
have access to a subset of the memory mapped file.
For now, only the `timestamp` module uses it.

Multiple threads _can not_ write to the same piece
of memory (obviously) so writes need to guarantee
one writer. For now this is done with a Mutex.

This adds a dependency on a new crate, `memmap`.