Problem: Windows clock is failing HLC guarantees
The Windows clock has a low resolution (100ns)
causing the HLC guarantees to fail during the
doctests. This was discovered after the removal of
static since a new clock was generated between
each test in the new implementation versus a
global clock before.

Solution: This is a temporary patch pending fixes
in upstream libraries to the Windows clock
resolution. For now a global clock is created for
the doctests and it is passed to the eval
function. This is not ideal.