Problem: Transaction logic is not contained
The transaction logic is spread out over a few
modules. This is the start of isolating it to
a single module where it can be tested in
isolation.

Solution: Move more of the transaction specific
logic into `txn_stack`. Also enforce more when it
comes to the type `TxnStack` can carry. I kept it
as generic to simplify testing.