Problem: messaging is only available through the server
It is currently impossible to do any messaging-related functionality via
pumpkindb_engine because SUBSCRIBE and UNSUBSCRIBE are essentially shims
implemented by the server.

This severely limits embedding scenario as it would only have access to the
post-execution stack but won't be able to send data during script execution
(useful in cursor iteration, etc.)

Solution: make messaging and pub/sub implementation pluggable and
adjust the server to use this mechanism.

This change renames SEND to PUBLISH and makes SUBSCRIBE push
subscription identifier back onto the top of the stack.

Closes #212