Problem: TIMES is quite slow
Current benchmark for 1000 DROPs done in TIMES loop is
right above 1ms on my laptop. That's pretty damn slow.

Solution: TIMES is doing a lot of allocation. Instead of
using that memory (which can grow significantly on larger
loops), simply push the reference to code N times to the
program stack. While it still uses memory, it is using
significantly less of it (as opposed to actually copying
the program over and over). This gives us approximately
20% benchmark improvement.