Problem: dispatching logic is intertwined with the Scheduler
This means the scheduler has to known which parameters to pass
to specific modules, which wouldn't work well if we're removing
or adding some modules.

Solution: replace Module trait with Dispatcher (of the same interface)
and extract current logic into StandardDispatcher. What's significant
about this is that now we're only passing relevant parameters
to the dispatcher, but not the scheduler as we're only providing
the final dispatcher to the scheduler.