Problem: new users have a barrier to try latest PumpkinDB out
They are required to prepare Rust development environment,
which is definitely a bit of a barrier.

Solution: provide a Dockerfile to build a docker image.

Note that in this version, for a yet unknown reason, when
`docker run -ti pumpkindb pumpkindb-term 172.17.0.1:9981`
is invoked, it messes up the terminal a little bit but
works fine after that. This is something we need to fix.