Problem: a number of doctests are being ignored
Running doctests skips over a number of doctests that are
defined, simply ignoring their existence.

Solution: the regular expression that was scanning for blocks
of tests did not account for empty lines.

Once this was addresses, it was discovered that a number of tests
actually fail, and this has been remedied as well.