Problem: writing custom servers is difficult
Basically, it requires duplicating a lot of code that is currently
in pumpkindb_server (with respect to mio and putting all the necessary
engine elements together)

Writing such servers is very useful when one needs to add new
built-in functionality to PumpkinDB yet retain its standard communication
protocol.

Solution: extract pumpkindb_mio_server crate that focuses primarily
on the mio part) and let pumpkindb_server handle the rest.

Note that this is still requiring one to at least look at pumpkindb_server
to implement their own custom server, but this is far better than before.

Also, the separation is not very clean, but this will start the ball
rolling.