Problem: no way to identify transaction boundaries
In certain cases it is worthwhile to learn whether some data was recorded
within the boundaries of the same transaction as other.

Solution: introduce TXID instruction that allows to do this by retrieving
ongoing transaction's ID. Transaction identifiers are guaranteed to be unique
and grow monotonically, thus allowing one to compare the order of their origination.