Problem: instructions handling macros are obscure
Specifically, `try_instruction!` and `instruction_is!` do not
communicate the fact that, given certain conditions (result-worthy
instruction invocaton or matching instruction name) they will
`return` a PassResult from the function that uses the macro.

This makes it harder to understand the flow of the code, unless
you're intimately familiar with these macros.

Solution: use `Try` trait (via `InstructionIs` struct -- arguably,
not the best name, please submit a PR if you can think of a better one)
and `?` syntax to indicate that in case of an error, the function
will be terminated and the error will be returned.

Also, add a `TryInstruction` trait that allows to build a chain
of handlers for instructions that couldn't be handled. These chains
will communicate the flow of the code more naturally.