Problem: in some cases, operating on stack triplets is useful
For example, I am currently working on ViewDB and my core
attribute processing instruction (`ATTR`) takes three elements
off the stack (id, attribute, value) and needs to pass them
to another instruction before operating on them itself.

Solution: implement `3DUP` and `3DROP` instructions that are
found in some Forth extensions. The good thing is that they
are following the same convention as `2DUP` and `2DROP` so it
is easy to guess what they do.