Problem: `TXID` manual doesn't specify identifier's length
Solution: specifying a specific length is potentially problematic.
Currently we use `HLC` to fulfill the obligations of `TXID`, but
this can eventually change (or not, hard to predict!).

However, it is important to provide certain guarantees about it. In
order to make all transaction identifiers perfectly comparable, they
are guaranteed to have a constant size.

This has been clarified in the documentation.