Problem: sharing stack with closures prevents typing
I've attempted to develop a simple typing inference mechanism for PumpkinScript
(https://github.com/yrashk/PumpkinDB/commit/d6c1e4617ac59575de65ddcebc630bac5d57c773)
but our closure evaluation rules prevent me from finishing it.

Basically, closures consume from the stack and push to the stack, and since we
don't know how many times a closure will be executed (can be anywhere between
zero and `n`), we don't really know how to infer or check type of subsequent
instructions.

Solution: instructions that evaluate closures a number of times that
is unknown compile-time have to evaluate these closures on a separate stack.

By only allowing such closures to communicate with the rest of the program
in a linear way, we can infer types. Also, it seems like a much cleaner
approach as well.

What has been done to implement this:

1. A concept of a stack of stacks was added, with corresponding "push" (`<`) and
"pop" (`>') instructions. This syntax was chosen to serve as a visual indicator of
stack separation:

    1 < ... > DUP

These instructions are mainly intended for internal use but because sometimes they
can be a great mechanism for other problems, they were made public.

This allows us to run such closures in new stacks and return to previous stacks.

2. In order to allow such closures to communicate return values linearly, Forth's
concept of a return stack with corresponding `>R` ("to R") and `R>` ("from R") instructions
were implemented. By only allowing to read one value at a time, we can infer the value type
by value's subsequent use (or type it through casting)

Note that the properties/guidelines for this return stack are different from Forth's,
as its primary motivation is communicating across "boundaries". For that reason,
if a better name will be found, it might be renamed to avoid the confusion.

3. The introduction of multiple stacks exacerbated dubious stack vector pre-allocation
technique, instructions like `TIMES` were getting really slow. By removing this technique,
we've simplified code and made new stack allocation much faster.

THIS IS A BREAKING CHANGE.

Closes #339