Problem: use of return stack to accumulate values reverses them
As of #340 we now must use >R and R> to pass the data out of a DOWHILE-like
loop. It's a working solution, however, it also means that if one to process
the return stack later on (in another loop), the values will be picked in a LIFO order.

Not a huge problem by itself, but in PumpkinDB, the loops are often used to
iterate over sorted keys and this means that the results of such "scans" will
be in a reverse order.

Solution: replace return stack with a (double-ended) queue.

This replaces `>R` and `R>` to `>Q` and `Q>` ("to the back of the queue"
and "from the back of the queue", respectively) and adds `<Q` and `Q<`
("to the front of the queue" and "from the front of the queue", respectively)

With this change it is possible to collect and process results in the same
order.

Closes #341
Closes #342