Problem: PumpkinDB is not *really* an event sourcing database engine
Its scope is much broader and, in a sense, this definition is a bit
misleading.

We have had a number of conversations where we agreed that "event sourcing"
can be a good way to boost interest.

But accuracy is important, too.

Solution: define PumpkinDB as an immutable ordered key-value database engine.

Also, still keep references to event sourcing in the README.